Minecraft 1.12 support and major changes
Update protocol to v335 (MCProtocolLib)
Replace various implementations with updated ones from MCProtocolLib
Remove obsolete modules:
*  GUI
* SpamBot
* Old protocols
Implement Login command

+++ core/src/main/java/org/darkstorm/minecraft/darkbot/ai/DerpTask.java
@@ -33,8 +33,8 @@ public class DerpTask extends AbstractTask {
 	@Override
 	public void run() {
 		MainPlayerEntity player = bot.getPlayer();
-		player.setYaw(random.nextDouble() * 360);
-		player.setPitch(random.nextDouble() * 360);
+		player.setYaw((float) random.nextDouble() * 360);
+		player.setPitch((float) random.nextDouble() * 360);
 	}
 
 	@Override
